backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
chromadb==0.5.15
PyPDF2==3.0.1
python-multipart==0.0.12
slowapi==0.1.9
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
